refactor(assistant-list): rename assistant list v2 implementation
