+++ src/renderer/src/components/chat/resources/variants/index.ts
@@ -1,11 +1,11 @@
 export { AgentResourceList } from './AgentList'
-export { AssistantResourceList } from './AssistantList'
+export { AssistantList, type AssistantListLabels, type AssistantListProps } from './AssistantList'
 export {
   type AssistantListActionContext,
   type AssistantListActionHandlers,
   createAssistantListActionRegistry
 } from './assistantListActions'
-export { AssistantListV2, type AssistantListV2Labels, type AssistantListV2Props } from './AssistantListV2'
+export { AssistantResourceList } from './AssistantResourceList'
 export { HistoryResourceList } from './HistoryList'
 export { SessionResourceList } from './SessionList'
 export { TopicResourceList } from './TopicList'
